refactor(rpc): move the server machinery into its own module
* `invenio_app/rpc.py` now holds the protocol and server plumbing
  (`recv_request`, `RPCRequestHandler`, `RPCServer`, `send_request`,
  `ServerStop`) with the click CLI injected into `RPCServer`, so the
  module depends on neither click nor flask.
* The `rpc-server` command group stays in `cli.py` next to the rest
  of the CLI. No behavior change.
